backend/requirements.txt
Django==4.1.1
django-cors-headers==3.12.0
django-debug-toolbar==3.6.0
django-rest-swagger==2.2.0
django-seed==0.3.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
python-decouple==3.6
psycopg2-binary==2.9.3
drf-access-policy==1.1.1
drf-yasg==1.20.0
simplejson==3.17.6
django-filter==22.1
